+++ src/components/common/Navbar/Navbar.tsx
@@ -9,7 +9,7 @@ import { List as ListIcon, X as XIcon } from 'phosphor-react'
 import SignInLink from 'src/components/ui/SignInLink'
 import SlideOver from 'src/components/ui/SlideOver'
 import useWindowDimensions from 'src/hooks/useWindowDimensions'
-import { useStoreCollection } from 'src/hooks/useStoreCollection'
+import { useStoreCollection } from 'src/hooks/useAllCollections'
 
 import SearchInput from '../SearchInput'
 
@@ -23,10 +23,10 @@ function NavLinks() {
   return (
     <nav className="navlinks__list">
       <UIList>
-        {links.map((item) => (
-          <li key={item.name}>
-            <Link variant="display" key={item.href} href={item.href}>
-              {item.name}
+        {links.map(({ node: link }) => (
+          <li key={link.seo.title}>
+            <Link variant="display" href={`/${link.slug}`}>
+              {link.seo.title}
             </Link>
           </li>
         ))}